refactor external compositiob service and added better documentation
