+++ packages/services/external-composition/federation-2/package.json
@@ -8,11 +8,11 @@
     "dev": "tsup-node src/dev.ts --format esm --shims --target node16 --watch --sourcemap --onSuccess 'node --enable-source-maps dist/dev.js' | pino-pretty --translateTime HH:MM:ss TT --ignore pid,hostname"
   },
   "dependencies": {
-    "@apollo/composition": "^2.2.1",
+    "@apollo/composition": "^2.2.2",
     "@whatwg-node/fetch": "^0.5.3",
     "@whatwg-node/server": "^0.4.17",
     "graphql": "^16.6.0",
-    "zod": "^3.15.1"
+    "zod": "^3.20.2"
   },
   "devDependencies": {
     "@graphql-hive/external-composition": "workspace:*"
