+++ packages/web/docs/src/pages/features/external-schema-composition.mdx
@@ -240,6 +240,40 @@ await server.listen({
 
 ### Apollo Federation v2 example
 
+#### CloudFlare Worker
+
+We provide a ready-to-use CloudFlare Worker project
+
+#### Pre-built Docker image
+
+We provide a
+[Docker image](https://github.com/kamilkisiela/graphql-hive/pkgs/container/graphql-hive%2Fcomposition-federation-2)
+for running external composition service for Apollo Federation v2.
+
+The pre-built image implements the best-practice to secure your endpoint, and uses the latest
+version of Apollo Federation v2.
+
+Start by deciding on your encryption secret. This is needed in order to ensure you endpoint is
+secured and can be triggered only by Hive platform. Your secret can be any string you decide, and it
+will be used as private key to hash the requests to your composition service.
+
+To run the container, you can use the following command:
+
+```
+docker run -p 3069 -e SECRET="MY_SECRET_HERE" ghcr.io/kamilkisiela/graphql-hive/composition-federation-2
+```
+
+You should make this service publicly available, and then configure it in Hive platform (see
+`Configuration` section above).
+
+The container created here listens to `POST /compose` requests, so the endpoint you are using should
+be composed from the public endpoint of this service, and the `/compose` path.
+
+#### Custom NodeJS Server
+
+You can also build your own server for the composition endpoint. The following example shows how to
+do that.
+
 The following example shows how to implement an external composition endpoint for Apollo Federation
 v2 in NodeJS.
